+++ Sources/Draw/GLRadiosityRenderer.cpp
@@ -109,7 +109,7 @@ namespace spades {
 						c.cz = z;
 					}
 
-			SPLog("Chunk buffer allocated (%d bytes)", sizeof(Chunk) * chunkW * chunkH * chunkD);
+			SPLog("Chunk buffer allocated (%d bytes)", (int) sizeof(Chunk) * chunkW * chunkH * chunkD);
 
 			// make texture
 			textureFlat = device->GenTexture();
